requirements.txt
opencv-python
numpy
pupil-labs_realtime_api
